backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asgi-lifespan>=2.1.0",
    "asgiref>=3.11.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "celery>=5.6.2",
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "scalar-fastapi>=1.5.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
